pyproject.toml
[project]
name = "colqwen-inference"
version = "0.1.0"
description = "ColQwen model inference service for text and image embeddings"
readme = "README.md"
requires-python = ">=3.12, <3.13"
dependencies = [
    "aiohttp>=3.9.0",
    "fastapi[standard]>=0.115.6",
    "modal>=0.71.3",
    "numpy>=1.24.0",
    "pydantic>=2.10.4",
    "python-dotenv>=1.0.1",
    "tenacity>=8.2.0",
    "tqdm>=4.67.1",
    "pillow",
]
[tool.uv.sources]
torch = [
    { index = "pytorch-cpu", marker = "platform_system == 'Linux' and platform_machine != 'aarch64'" },
]
torchvision = [
    { index = "pytorch-cpu", marker = "platform_system == 'Linux' and platform_machine != 'aarch64'" },

]

# https://docs.astral.sh/uv/guides/integration/pytorch/#using-a-pytorch-index
# https://github.com/astral-sh/uv/issues/8358#issuecomment-2424808369
[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[dependency-groups]
dev = ["ipykernel>=6.29.5"]
